DRY datetime validation and sim test helpers
