+++ packages/doeff-time/src/doeff_time/_internals/time_queue.py
@@ -8,13 +8,7 @@ from typing import Any
 
 from doeff.effects.spawn import Promise
 
-
-def _ensure_aware_datetime(value: datetime, *, name: str) -> datetime:
-    if not isinstance(value, datetime):
-        raise TypeError(f"{name} must be datetime, got {type(value).__name__}")
-    if value.tzinfo is None:
-        raise ValueError(f"{name} must be timezone-aware datetime")
-    return value
+from .validation import ensure_aware_datetime
 
 
 @dataclass(frozen=True)
@@ -30,7 +24,7 @@ class TimeQueue:
         self._items: list[tuple[datetime, int, Promise[Any]]] = []
 
     def push(self, time: datetime, promise: Promise[Any]) -> None:
-        target_time = _ensure_aware_datetime(time, name="time")
+        target_time = ensure_aware_datetime(time, name="time")
         self._sequence += 1
         heapq.heappush(self._items, (target_time, self._sequence, promise))
