+++ packages/doeff-time/tests/test_sim_time.py
@@ -4,6 +4,7 @@ from dataclasses import dataclass
 from datetime import datetime, timedelta, timezone
 from typing import Any
 
+from conftest import SIM_TIME_EPOCH, sim_seconds, sim_time
 from doeff_events import Publish, WaitForEvent, event_handler
 from doeff_time import Delay, GetTime, ScheduleAt, SetTime, WaitUntil, sim_time_handler
 
@@ -22,16 +23,6 @@ from doeff import (
 )
 from doeff.effects.gather import GatherEffect
 from doeff.effects.spawn import PRIORITY_HIGH, SpawnEffect
-
-SIM_TIME_EPOCH = datetime(2024, 1, 1, tzinfo=timezone.utc)
-
-
-def sim_time(seconds: float) -> datetime:
-    return SIM_TIME_EPOCH + timedelta(seconds=seconds)
-
-
-def sim_seconds(value: datetime) -> float:
-    return (value - SIM_TIME_EPOCH).total_seconds()
 
 
 def _run_with_sim(
